+++ sdk/monitor/azure-monitor-opentelemetry/CHANGELOG.md
@@ -9,6 +9,8 @@
 
 - Add ability to specify which logger to export telemetry for via `logger_name` configuration
     ([#32192](https://github.com/Azure/azure-sdk-for-python/pull/32192))
+- Allow OTEL_PYTHON_DISABLED_INSTRUMENTATIONS functionality for Azure Core Tracing in Auto-instrumentation
+    ([#XXXXX](https://github.com/Azure/azure-sdk-for-python/pull/XXXXX))
 
 ### Breaking Changes